split Custom into CustomRegressor and CustomClassifier
