+++ nnetsauce/__init__.py
@@ -1,17 +1,9 @@
 from .base import Base 
-from .custom import Custom
+from .custom.customClassifier import CustomClassifier
+from .custom.customRegressor import CustomRegressor
 from .mts import MTS
 from .rvfl.bayesianrvfl import BayesianRVFL
 from .rvfl.bayesianrvfl2 import BayesianRVFL2 
 
 
-__all__=["Base", "BayesianRVFL", "BayesianRVFL2", "Custom", "MTS"]
-
-
-#from . import base
-#from . import custom
-#from . import simulation
-#from . import utils
-
-
-#__all__=["base", "custom", "simulation", "utils"]
+__all__=["Base", "BayesianRVFL", "BayesianRVFL2", "CustomClassifier", "CustomRegressor", "MTS"]
