+++ nnetsauce/custom/__init__.py
@@ -1,3 +1,4 @@
-from .custom import Custom
+from .customClassifier import CustomClassifier
+from .customRegressor import CustomRegressor
 
-__all__ = ["Custom"]
+__all__ = ["CustomClassifier", "CustomRegressor"]
